Add files via upload
 Magic Number Game
In this fun scenario, you can create a game where the user has to guess a "magic number" generated by the program. The twist is that the user gets clues after each guess, and the program uses a user-defined function to provide hints.

Problem Statement:

The program generates a random "magic number" between 1 and 100.
The user will try to guess the number.
After each guess, the program will provide a hint ("Too high", "Too low", or "Correct").
The game continues until the user guesses correctly.
Use a user-defined function to evaluate and provide the hint.
User-defined function:

void provideHint(int guess, int magicNumber);
This function will compare the guessed number with the magic number and print an appropriate message.