Add files via upload
 Day of the Week Activity Suggestion
Description: Write a C program that suggests an activity based on the day of the week. The program should take an integer input from the user (1 for Monday, 2 for Tuesday, ..., 7 for Sunday) and use a switch case statement to output a suggested activity for that day. If the input does not correspond to a valid day of the week, the program should print an error message.

Example Activities:

Monday (1): "Start the week with some exercise!"
Tuesday (2): "Read a new book."
Wednesday (3): "Mid-week movie night!"
Thursday (4): "Try cooking a new recipe."
Friday (5): "Hang out with friends."
Saturday (6): "Go for a hike or outdoor activity."
Sunday (7): "Relax and prepare for the week ahead."
Example Input/Output:

Input: 3
Output: Mid-week movie night!
Input: 7
Output: Relax and prepare for the week ahead.
Input: 8
Output: Error: Invalid day. Please enter a number between 1 and 7.